Move interpolate to timeseries scale
